+++ MyApp/src/Application/Modules/BackgroundJobs/BaseBackgroundJob.cs
@@ -1,6 +1,6 @@
 ﻿using Quartz;
 
-namespace MyApp.Application.Infrastructure;
+namespace MyApp.Application.Modules.BackgroundJobs;
 public abstract class BaseBackgroundJob<TActual> : IJob
 {
     public static string Name { get; } = typeof(TActual).Name;
