+++ MyApp/src/Application/Startup/Behaviours/ValidationBehavior.cs
@@ -4,7 +4,7 @@ using MediatR;
 
 namespace MyApp.Application.Startup.Behaviours;
 
-public class ValidationBehaviour<TRequest, TResponse>(
+public class ValidationBehavior<TRequest, TResponse>(
     IEnumerable<IValidator<TRequest>> validators
     ) : IPipelineBehavior<TRequest, TResponse>
     where TRequest : notnull
