+++ MyApp/src/Infrastructure/BackgroundJobs/BackgroundJobsStarter.cs
@@ -1,4 +1,4 @@
-﻿using MyApp.Application.Infrastructure;
+﻿using MyApp.Application.Modules.BackgroundJobs;
 using Quartz;
 
 namespace MyApp.Infrastructure.BackgroundJobs;
