+++ MyApp/tests/Tests.Integration/Tests/Features/Cleanup/CleanupConfirmationsTests.cs
@@ -1,5 +1,4 @@
-﻿using MyApp.Application.Infrastructure.BackgroundJobs.Cleanup.Confirmations;
-using MyApp.Domain.Auth.EmailChangeConfirmation;
+﻿using MyApp.Domain.Auth.EmailChangeConfirmation;
 using MyApp.Domain.Auth.PasswordResetConfirmation;
 using MyApp.Domain.Auth.UserConfirmation;
 using MyApp.Domain.Shared;
