+++ MyApp/tests/Tests.Integration/Tests/Features/Cleanup/CleanupInstancesTests.cs
@@ -1,5 +1,4 @@
-﻿using MyApp.Application.Infrastructure;
-using MyApp.Application.Infrastructure.BackgroundJobs.Cleanup.Instances;
+﻿using MyApp.Application.Infrastructure.Repositories;
 using MyApp.Domain.Infra.Instance;
 
 namespace MyApp.Tests.Integration.Tests.Features.Cleanup;
