+++ MyApp/tests/Tests.System/Providers/Email/EmailProvider.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using MailKit;
+﻿using MailKit;
 using MailKit.Net.Imap;
 using MailKit.Search;
 using MimeKit;
